// src/types/strapi.d.ts
export interface components {
    schemas: {
// ...
        DynamicZoneDoctorStoriesComponent: {
            id?: number;
            /** @enum {string} */
            __component?: "dynamic-zone.doctor-stories";
            highlighted_title_line1?: string;
            highlighted_title_line2?: string;
            title?: string;
            stories?: components["schemas"]["ItemsUserstoryCardComponent"][];
        };
// ...
    };
// ...
}